refactor and chores
