+++ bin/eckode.mjs
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-import '../src/index.mjs'
+import '../src/index.mjs';
