+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eckode/cli",
-  "version": "0.0.5",
+  "version": "0.0.7",
   "description": "CLI tool, written in TypeScript, for running @eckode scoped packages.",
   "keywords": [
     "eckode",
@@ -20,16 +20,16 @@
   },
   "homepage": "https://github.com/eckode/cli",
   "scripts": {
-    "build": "yarn clean && tsc",
-    "clean": "ts-node-esm ./src/clean.mts",
-    "pkg": "tsc && ts-node-esm ./src/package.mts"
+    "build": "npm run clean && tsc",
+    "clean": "ts-node-esm ./src/clean.mjs",
+    "pkg": "tsc && ts-node-esm ./src/package.mjs"
   },
   "bin": {
     "eckode": "./bin/eckode.mjs"
   },
   "engines": {
-    "node": ">=16",
-    "npm": ">=8"
+    "node": ">=18",
+    "npm": ">=9"
   },
   "files": [
     "bin",
@@ -43,25 +43,23 @@
     "access": "public"
   },
   "devDependencies": {
-    "@types/express": "^4.17.17",
-    "@types/node": "^18.15.13",
-    "@types/resolve-bin": "^0.4.1",
-    "@typescript-eslint/eslint-plugin": "^5.59.0",
-    "@typescript-eslint/parser": "^5.59.0",
-    "chalk": "^5.2.0",
-    "eslint": "^8.39.0",
-    "eslint-config-standard-with-typescript": "^34.0.1",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-n": "^15.7.0",
-    "eslint-plugin-promise": "^6.1.1",
-    "jest": "^29.5.0",
-    "prettier": "^2.8.7",
-    "resolve-bin": "^1.0.1",
-    "ts-node": "^10.9.1",
-    "typescript": "~5.0.4"
+    "@types/express": "^5.0.0",
+    "@types/node": "^22.10.1",
+    "@types/resolve-bin": "^0.4.3",
+    "@typescript-eslint/eslint-plugin": "^8.18.0",
+    "@typescript-eslint/parser": "^8.18.0",
+    "chalk": "^5.3.0",
+    "eslint": "^9.16.0",
+    "eslint-config-love": "^110.0.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-n": "^17.15.0",
+    "eslint-plugin-promise": "^7.2.1",
+    "jest": "^29.7.0",
+    "prettier": "^3.4.2",
+    "rimraf": "^6.0.1",
+    "ts-node": "^10.9.2",
+    "typescript": "~5.7.2"
   },
   "main": "./src/index.mjs",
-  "dependencies": {
-    "rimraf": "^5.0.0"
-  }
+  "dependencies": {}
 }
