+++ src/index.mts
@@ -3,7 +3,7 @@ import { exit } from 'process';
 
 import { spawnNodeProcess, logger } from './utils.mjs';
 import nodeLocalVars from './node-env-vars.mjs';
-import { EckodeNpmScripts } from './types.js';
+import { EckodeNpmScripts } from './types.mjs';
 
 const eckodeNpmCommands: Array<EckodeNpmScripts> = ['build', 'dev'];
 
@@ -21,8 +21,7 @@ process.env.ECKO_PROJECT_PATH = process.cwd();
 process.env.ECKO_COMMAND = lifecycleEvent as EckodeNpmScripts;
 
 if (!eckodeNpmCommands.includes(process.env.ECKO_COMMAND)) {
-  logger(`command not found!`, 'error', process.env.ECKO_COMMAND),
-  exit(1);
+  logger(`command not found!`, 'error', process.env.ECKO_COMMAND), exit(1);
 }
 
 // Start er up!
